requirements.txt
pandas==3.0.0
Pillow==12.1.0
scikit_learn==1.8.0
streamlit==1.52.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
